jpa_updated

+++ src/main/java/guru/springframework/spring5webapp/domain/Author.java
@@ -1,41 +1,37 @@
 package guru.springframework.spring5webapp.domain;
 
 import java.util.Set;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToMany;
+import javax.persistence.*;
 
 @Entity
 public class Author {
 
 	@Id
-	@GeneratedValue(strategy=GenerationType.AUTO)
-	private int Id;
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	private Long Id;
 
 	private String firstName;
 	private String lastName;
 	
-	@ManyToMany(mappedBy="authors")
+	@ManyToMany(mappedBy = "authors")
 	private Set<Book> books;
 
 	public Author() {
 		
 	}
 
-	public Author(int id, String firstName, String lastName, Set<Book> books) {
+	public Author(Long  id, String firstName, String lastName, Set<Book> books) {
 		Id = id;
 		this.firstName = firstName;
 		this.lastName = lastName;
 		this.books = books;
 	}
 
-	public int getId() {
+	public Long  getId() {
 		return Id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long  id) {
 		Id = id;
 	}
